Remove gamepad translations
